Refactor: common 패키지의 domain 패키지 네임을 entity로 변경(BaseEntity)
